feat(faq): extract tab content
